Save instrument refactor
